Moved Helper functionality related to immutability-checking to ImmutableTypes class
